replacing ch.vorburger.hotea.minecraft.api design (see details)
In the v1 & v2 design this was an API for "hot" plugins themselves (each
one); whereas in v3 this is just a glue with the HotPluginManager
service interface for the (one and only) ch.vorburger.hotea.minecraft
plugin to talk to the patched Sponge.  The (individual) "hot" plugins
are now "normal" Sponge plugins, with nothing in particular to them
anymore.